fix(pro): cleanup local workspaces if pro instance is deleted before shutting down daemon and removing provider from config
