+++ cmd/provider/delete.go
@@ -4,15 +4,12 @@ import (
 	"context"
 	"fmt"
 	"os"
-	"sync"
 
 	"github.com/loft-sh/devpod/cmd/flags"
-	"github.com/loft-sh/devpod/pkg/client/clientimplementation"
 	"github.com/loft-sh/devpod/pkg/config"
 	provider2 "github.com/loft-sh/devpod/pkg/provider"
 	"github.com/loft-sh/devpod/pkg/workspace"
 	logpkg "github.com/loft-sh/log"
-	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 )
 
@@ -69,7 +66,7 @@ func (cmd *DeleteCmd) Run(ctx context.Context, args []string) error {
 	}
 
 	// delete the provider
-	err = DeleteProvider(ctx, devPodConfig, provider, cmd.IgnoreNotFound, false, logpkg.Default)
+	err = DeleteProvider(ctx, devPodConfig, provider, cmd.IgnoreNotFound, logpkg.Default)
 	if err != nil {
 		return err
 	}
@@ -78,57 +75,31 @@ func (cmd *DeleteCmd) Run(ctx context.Context, args []string) error {
 	return nil
 }
 
-func DeleteProvider(ctx context.Context, devPodConfig *config.Config, provider string, ignoreNotFound bool, cleanup bool, log logpkg.Logger) error {
+func DeleteProvider(ctx context.Context, devPodConfig *config.Config, provider string, ignoreNotFound bool, log logpkg.Logger) error {
 	// check if there are workspaces that still use this provider
 	workspaces, err := workspace.List(ctx, devPodConfig, true, log)
 	if err != nil {
 		return err
 	}
-	usedWorkspaces := []*provider2.Workspace{}
+
+	// search for workspace that uses this machine
 	for _, workspace := range workspaces {
 		if workspace.Provider.Name == provider {
-			usedWorkspaces = append(usedWorkspaces, workspace)
-		}
-	}
-
-	if len(usedWorkspaces) > 0 {
-		if cleanup {
-			wg := sync.WaitGroup{}
-			// try to force delete all workspaces in the background
-			for _, w := range usedWorkspaces {
-				wg.Add(1)
-				go func(w provider2.Workspace) {
-					defer wg.Done()
-					client, err := workspace.Get(ctx, devPodConfig, []string{w.ID}, false, log)
-					if err != nil {
-						log.Errorf("Failed to get workspace %s: %v", w.ID, err)
-						return
-					}
-					// delete workspace folder
-					err = clientimplementation.DeleteWorkspaceFolder(devPodConfig.DefaultContext, client.Workspace(), client.WorkspaceConfig().SSHConfigPath, log)
-					if err != nil {
-						log.Errorf("Failed to delete workspace %s: %v", w.ID, err)
-						return
-					}
-					log.Donef("Successfully deleted workspace %s", w.ID)
-				}(*w)
-			}
-
-			log.Infof("Waiting for %d workspaces to be deleted", len(usedWorkspaces))
-			wg.Wait()
-		} else {
-			workspace := usedWorkspaces[0]
 			return fmt.Errorf("cannot delete provider '%s', because workspace '%s' is still using it. Please delete the workspace '%s' before deleting the provider", workspace.Provider.Name, workspace.ID, workspace.ID)
 		}
 	}
 
+	return DeleteProviderConfig(devPodConfig, provider, ignoreNotFound)
+}
+
+func DeleteProviderConfig(devPodConfig *config.Config, provider string, ignoreNotFound bool) error {
 	if devPodConfig.Current().DefaultProvider == provider {
 		devPodConfig.Current().DefaultProvider = ""
 	}
 	delete(devPodConfig.Current().Providers, provider)
-	err = config.SaveConfig(devPodConfig)
+	err := config.SaveConfig(devPodConfig)
 	if err != nil {
-		return errors.Wrap(err, "save config")
+		return fmt.Errorf("save config: %w", err)
 	}
 
 	providerDir, err := provider2.GetProviderDir(devPodConfig.DefaultContext, provider)
@@ -149,7 +120,7 @@ func DeleteProvider(ctx context.Context, devPodConfig *config.Config, provider s
 	}
 	err = os.RemoveAll(providerDir)
 	if err != nil {
-		return errors.Wrap(err, "delete provider dir")
+		return fmt.Errorf("delete provider dir: %w", err)
 	}
 
 	return nil
