+++ pkg/daemon/platform/local_server.go
@@ -226,8 +226,8 @@ func (l *localServer) metrics(w http.ResponseWriter, r *http.Request, params htt
 
 func (l *localServer) shutdown(w http.ResponseWriter, r *http.Request, params httprouter.Params) {
 	err := l.Close()
-	if err == nil {
-		http.Error(w, fmt.Errorf("shut down daemon server: %w", err).Error(), http.StatusInternalServerError)
+	if err != nil {
+		http.Error(w, fmt.Errorf("shut down daemon server: %v", err).Error(), http.StatusInternalServerError)
 		return
 	}
